# apps/push-click-service/src/main/java/io/clickhouselab/pushclick/web/dto/RecordClickResponse.java
package io.clickhouselab.pushclick.web.dto;

import io.swagger.v3.oas.annotations.media.Schema;

import java.util.UUID;

@Schema(description = "Confirmation that a click event was accepted.")
public record RecordClickResponse(

        @Schema(description = "Server-generated identifier for this click event.")
        UUID clickId
) {
}